backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
# psycopg2-binary removed — not needed, we use asyncpg
alembic==1.14.1

# Redis
redis==5.2.1

# Vector Database
chromadb>=0.6.3

# NumPy
#numpy==1.26.4

# LangChain & LangGraph
langchain==0.3.14
langchain-groq>=0.2.0
langchain-ollama>=0.2.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
langchain-community==0.3.14
langchain-chroma==0.2.2
langgraph==0.2.61

# (Google Gemini removed — using Groq + Ollama + HuggingFace instead)

# Git operations
gitpython==3.1.44

# Streaming
sse-starlette==2.2.1

# Code parsing
tree-sitter==0.24.0

# Utilities
aiofiles==24.1.0
httpx==0.28.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pydantic[email]==2.10.4
